+++ src/main/java/com/sd64/novastore/controller/admin/BillController.java
@@ -1,12 +1,13 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
-import com.sd64.novastore.request.AccountRequest;
-import com.sd64.novastore.service.AccountService;
+import com.sd64.novastore.request.BillRequest;
+import com.sd64.novastore.service.BillService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.ObjectError;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,50 +19,44 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-
 @RestController
-@RequestMapping("/account")
-public class AccountController {
+@RequestMapping("/bill")
+public class BillController {
     @Autowired
-    private AccountService accountService;
+    private BillService billService;
 
-    @GetMapping("/all")
+    @GetMapping("/getall")
     public ResponseEntity<?> getAll() {
-        return ResponseEntity.ok(accountService.getAll());
+        return ResponseEntity.ok(billService.getAll());
     }
 
-    @GetMapping("/page")
+    @GetMapping("/getallpt")
     public ResponseEntity<?> getAllPT(@RequestParam(defaultValue = "0", value = "page") Integer page) {
-        return ResponseEntity.ok(accountService.getAllPT(page).getContent());
+        return ResponseEntity.ok(billService.getAllPT(page).getContent());
     }
 
-    @PutMapping("/delete/{id}")
+    @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> delete(@PathVariable Integer id) {
-        return ResponseEntity.ok(accountService.delete(id));
+        return ResponseEntity.ok(billService.delete(id));
     }
 
 
     @PostMapping("/add")
-    public ResponseEntity<?> add(@RequestBody @Valid AccountRequest accountRequest, BindingResult result) {
+    public ResponseEntity<?> add(@RequestBody @Valid BillRequest billRequest, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
         }
-        return ResponseEntity.ok(accountService.add(accountRequest));
+        return ResponseEntity.ok(billService.add(billRequest));
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> update(@RequestBody @Valid AccountRequest accountRequest, @PathVariable Integer id, BindingResult result) {
+    public ResponseEntity<?> update(@RequestBody @Valid BillRequest billRequest, @PathVariable Integer id, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
         }
-        return ResponseEntity.ok(accountService.update(accountRequest, id));
+        return ResponseEntity.ok(billService.update(billRequest, id));
 
-    }
-
-    @GetMapping("/search")
-    public ResponseEntity<?> search(@RequestParam String name, @RequestParam(defaultValue = "0") int page) {
-        return ResponseEntity.ok(accountService.search(name, page).getContent());
     }
 }
