+++ src/main/java/com/sd64/novastore/controller/admin/CartController.java
@@ -1,7 +1,7 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
-import com.sd64.novastore.request.BillHistoryRequest;
-import com.sd64.novastore.service.BillHistoryService;
+import com.sd64.novastore.request.CartRequest;
+import com.sd64.novastore.service.CartService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -20,43 +20,42 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/bill_history")
-public class BillHistoryController {
+@RequestMapping("/cart")
+public class CartController {
     @Autowired
-    private BillHistoryService billHistoryService;
+    private CartService cartService;
 
     @GetMapping("/getall")
     public ResponseEntity<?> getAll() {
-        return ResponseEntity.ok(billHistoryService.getAll());
+        return ResponseEntity.ok(cartService.getAll());
     }
 
     @GetMapping("/getallpt")
     public ResponseEntity<?> getAllPT(@RequestParam(defaultValue = "0", value = "page") Integer page) {
-        return ResponseEntity.ok(billHistoryService.getAllPT(page).getContent());
+        return ResponseEntity.ok(cartService.getAllPT(page).getContent());
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> delete(@PathVariable Integer id) {
-        return ResponseEntity.ok(billHistoryService.delete(id));
+        return ResponseEntity.ok(cartService.delete(id));
     }
 
 
     @PostMapping("/add")
-    public ResponseEntity<?> add(@RequestBody @Valid BillHistoryRequest billHistoryRequest, BindingResult result) {
+    public ResponseEntity<?> add(@RequestBody @Valid CartRequest cartRequest, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
         }
-        return ResponseEntity.ok(billHistoryService.add(billHistoryRequest));
+        return ResponseEntity.ok(cartService.add(cartRequest));
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> update(@RequestBody @Valid BillHistoryRequest billHistoryRequest, @PathVariable Integer id, BindingResult result) {
+    public ResponseEntity<?> update(@RequestBody @Valid CartRequest cartRequest, @PathVariable Integer id, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
         }
-        return ResponseEntity.ok(billHistoryService.update(billHistoryRequest, id));
-
+        return ResponseEntity.ok(cartService.update(cartRequest, id));
     }
 }
