+++ src/main/java/com/sd64/novastore/controller/admin/CartDetailController.java
@@ -1,7 +1,7 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
-import com.sd64.novastore.request.AddressRequest;
-import com.sd64.novastore.service.AddressService;
+import com.sd64.novastore.request.CartDetailRequest;
+import com.sd64.novastore.service.CartDetailService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -20,43 +20,42 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/address")
-public class AddressController {
+@RequestMapping("/cart_detail")
+public class CartDetailController {
     @Autowired
-    private AddressService addressService;
+    private CartDetailService cartDetailService;
 
     @GetMapping("/getall")
     public ResponseEntity<?> getAll() {
-        return ResponseEntity.ok(addressService.getAll());
+        return ResponseEntity.ok(cartDetailService.getAll());
     }
 
     @GetMapping("/getallpt")
     public ResponseEntity<?> getAllPT(@RequestParam(defaultValue = "0", value = "page") Integer page) {
-        return ResponseEntity.ok(addressService.getAllPT(page).getContent());
+        return ResponseEntity.ok(cartDetailService.getAllPT(page).getContent());
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> delete(@PathVariable Integer id) {
-        return ResponseEntity.ok(addressService.delete(id));
+        return ResponseEntity.ok(cartDetailService.delete(id));
     }
 
 
     @PostMapping("/add")
-    public ResponseEntity<?> add(@RequestBody @Valid AddressRequest addressRequest, BindingResult result) {
+    public ResponseEntity<?> add(@RequestBody @Valid CartDetailRequest cartDetailRequest, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
         }
-        return ResponseEntity.ok(addressService.add(addressRequest));
+        return ResponseEntity.ok(cartDetailService.add(cartDetailRequest));
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> update(@RequestBody @Valid AddressRequest addressRequest, @PathVariable Integer id, BindingResult result) {
+    public ResponseEntity<?> update(@RequestBody @Valid CartDetailRequest cartDetailRequest, @PathVariable Integer id, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
         }
-        return ResponseEntity.ok(addressService.update(addressRequest, id));
-
+        return ResponseEntity.ok(cartDetailService.update(cartDetailRequest, id));
     }
 }
