+++ src/main/java/com/sd64/novastore/controller/admin/CategoryController.java
@@ -1,12 +1,13 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
-import com.sd64.novastore.request.FormRequest;
-import com.sd64.novastore.service.FormService;
+import com.sd64.novastore.request.CategoryRequest;
+import com.sd64.novastore.service.CategoryService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.ObjectError;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -19,52 +20,47 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/form")
-public class FormController {
+@RequestMapping("/category")
+public class CategoryController {
     @Autowired
-    private FormService formService;
+    private CategoryService categoryService;
 
     @GetMapping("/all")
     public ResponseEntity<?> getAll() {
-        return ResponseEntity.ok(formService.getAll());
+        return ResponseEntity.ok(categoryService.getAll());
     }
 
     @GetMapping("/page")
-    public ResponseEntity<?> getAll(@RequestParam(defaultValue = "0", name = "page") Integer page) {
-        return ResponseEntity.ok(formService.getAll(page).getContent());
+    public ResponseEntity<?> getAllPT(@RequestParam(defaultValue = "0", value = "page") Integer page) {
+        return ResponseEntity.ok(categoryService.getAllPT(page).getContent());
     }
 
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<?> delete(@PathVariable Integer id) {
+        return ResponseEntity.ok(categoryService.delete(id));
+    }
+
+
     @PostMapping("/add")
-    public ResponseEntity<?> add(@RequestBody @Valid FormRequest formRequest, BindingResult result) {
+    public ResponseEntity<?> add(@RequestBody @Valid CategoryRequest categoryRequest, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
-        } else {
-            return ResponseEntity.ok(formService.add(formRequest));
         }
+        return ResponseEntity.ok(categoryService.add(categoryRequest));
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> update(@PathVariable Integer id, @RequestBody @Valid FormRequest formRequest, BindingResult result) {
+    public ResponseEntity<?> update(@RequestBody @Valid CategoryRequest categoryRequest, @PathVariable Integer id, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
-        } else {
-            return ResponseEntity.ok(formService.update(formRequest, id));
         }
-    }
-
-    @PutMapping("/delete/{id}")
-    public ResponseEntity<?> delete(@PathVariable Integer id) {
-        if (formService.delete(id)) {
-            return ResponseEntity.ok("Delete Success");
-        } else {
-            return ResponseEntity.ok("Delete Fail");
-        }
+        return ResponseEntity.ok(categoryService.update(categoryRequest, id));
     }
 
     @GetMapping("/search")
     public ResponseEntity<?> search(@RequestParam String name, @RequestParam(defaultValue = "0") int page) {
-        return ResponseEntity.ok(formService.search(name, page).getContent());
+        return ResponseEntity.ok(categoryService.search(name, page).getContent());
     }
 }
