+++ src/main/java/com/sd64/novastore/controller/admin/FormController.java
@@ -1,13 +1,12 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
-import com.sd64.novastore.request.BrandRequest;
-import com.sd64.novastore.service.BrandService;
+import com.sd64.novastore.request.FormRequest;
+import com.sd64.novastore.service.FormService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.ObjectError;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,48 +17,54 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+
 @RestController
-@RequestMapping("/brand")
-public class BrandController {
+@RequestMapping("/form")
+public class FormController {
     @Autowired
-    private BrandService brandService;
+    private FormService formService;
 
     @GetMapping("/all")
     public ResponseEntity<?> getAll() {
-        return ResponseEntity.ok(brandService.getAll());
+        return ResponseEntity.ok(formService.getAll());
     }
 
     @GetMapping("/page")
-    public ResponseEntity<?> getAllPT(@RequestParam(defaultValue = "0", value = "page") Integer page) {
-        return ResponseEntity.ok(brandService.getAllPT(page).getContent());
+    public ResponseEntity<?> getAll(@RequestParam(defaultValue = "0", name = "page") Integer page) {
+        return ResponseEntity.ok(formService.getAll(page).getContent());
     }
-
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> delete(@PathVariable Integer id) {
-        return ResponseEntity.ok(brandService.delete(id));
-    }
-
 
     @PostMapping("/add")
-    public ResponseEntity<?> add(@RequestBody @Valid BrandRequest brandRequest, BindingResult result) {
+    public ResponseEntity<?> add(@RequestBody @Valid FormRequest formRequest, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
+        } else {
+            return ResponseEntity.ok(formService.add(formRequest));
         }
-        return ResponseEntity.ok(brandService.add(brandRequest));
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> update(@RequestBody @Valid BrandRequest brandRequest, @PathVariable Integer id, BindingResult result) {
+    public ResponseEntity<?> update(@PathVariable Integer id, @RequestBody @Valid FormRequest formRequest, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
+        } else {
+            return ResponseEntity.ok(formService.update(formRequest, id));
         }
-        return ResponseEntity.ok(brandService.update(brandRequest, id));
+    }
+
+    @PutMapping("/delete/{id}")
+    public ResponseEntity<?> delete(@PathVariable Integer id) {
+        if (formService.delete(id)) {
+            return ResponseEntity.ok("Delete Success");
+        } else {
+            return ResponseEntity.ok("Delete Fail");
+        }
     }
 
     @GetMapping("/search")
     public ResponseEntity<?> search(@RequestParam String name, @RequestParam(defaultValue = "0") int page) {
-        return ResponseEntity.ok(brandService.search(name, page).getContent());
+        return ResponseEntity.ok(formService.search(name, page).getContent());
     }
 }
