+++ src/main/java/com/sd64/novastore/controller/admin/MaterialController.java
@@ -1,4 +1,4 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
 import com.sd64.novastore.request.MaterialRequest;
 import com.sd64.novastore.service.MaterialService;
