+++ src/main/java/com/sd64/novastore/controller/admin/ProductController.java
@@ -1,4 +1,4 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
 import com.sd64.novastore.request.ProductRequest;
 import com.sd64.novastore.service.ProductService;
