+++ src/main/java/com/sd64/novastore/controller/admin/ProductDetailController.java
@@ -1,13 +1,12 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
-import com.sd64.novastore.request.ImageRequest;
-import com.sd64.novastore.service.ImageService;
+import com.sd64.novastore.request.ProductDetailRequest;
+import com.sd64.novastore.service.ProductDetailService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.ObjectError;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -20,44 +19,44 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/image")
-public class ImageController {
+@RequestMapping("/product-detail")
+public class ProductDetailController {
     @Autowired
-    private ImageService imageService;
+    private ProductDetailService productDetailService;
 
     @GetMapping("/all")
     public ResponseEntity<?> getAll() {
-        return ResponseEntity.ok(imageService.getAll());
+        return ResponseEntity.ok(productDetailService.getAll());
     }
 
     @GetMapping()
     public ResponseEntity<?> getAll(@RequestParam(defaultValue = "0", name = "page") Integer page) {
-        return ResponseEntity.ok(imageService.getAll(page).getContent());
+        return ResponseEntity.ok(productDetailService.getAll(page).getContent());
     }
 
     @PostMapping("/add")
-    public ResponseEntity<?> add(@RequestBody @Valid ImageRequest imageRequest, BindingResult result) {
+    public ResponseEntity<?> add(@RequestBody @Valid ProductDetailRequest productDetailRequest, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
         } else {
-            return ResponseEntity.ok(imageService.add(imageRequest));
+            return ResponseEntity.ok(productDetailService.add(productDetailRequest));
         }
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> update(@PathVariable Integer id, @RequestBody @Valid ImageRequest imageRequest, BindingResult result) {
+    public ResponseEntity<?> update(@PathVariable Integer id, @RequestBody @Valid ProductDetailRequest productDetailRequest, BindingResult result) {
         if (result.hasErrors()) {
             List<ObjectError> list = result.getAllErrors();
             return ResponseEntity.ok(list);
         } else {
-            return ResponseEntity.ok(imageService.update(imageRequest, id));
+            return ResponseEntity.ok(productDetailService.update(productDetailRequest, id));
         }
     }
 
-    @DeleteMapping("/delete/{id}")
+    @PutMapping("/delete/{id}")
     public ResponseEntity<?> delete(@PathVariable Integer id) {
-        if (imageService.delete(id)) {
+        if (productDetailService.delete(id)) {
             return ResponseEntity.ok("Xoá thành công");
         } else {
             return ResponseEntity.ok("Xoá thất bại");
