+++ src/main/java/com/sd64/novastore/controller/admin/PromotionController.java
@@ -1,9 +1,7 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
 import com.sd64.novastore.request.PromotionRequest;
-import com.sd64.novastore.request.SizeRequest;
 import com.sd64.novastore.service.PromotionService;
-import com.sd64.novastore.service.SizeService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -11,40 +9,40 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/size")
-public class SizeController {
+@RequestMapping("/promotion")
+public class PromotionController {
 
     @Autowired
-    private SizeService sizeService;
+    private PromotionService promotionService;
 
     @GetMapping("/all")
     public ResponseEntity<?> getAll() {
-        return ResponseEntity.ok(sizeService.getAll());
+        return ResponseEntity.ok(promotionService.getAll());
     }
 
     @GetMapping("/page")
     public ResponseEntity<?> getPage(@RequestParam(defaultValue = "0") int page) {
-        return ResponseEntity.ok(sizeService.getPage(page).getContent());
+        return ResponseEntity.ok(promotionService.getPage(page).getContent());
     }
 
     @PostMapping("/add")
-    public ResponseEntity<?> add(@RequestBody @Valid SizeRequest sizeRequest, BindingResult bindingResult) {
+    public ResponseEntity<?> add(@RequestBody @Valid PromotionRequest promotionRequest, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             String error = bindingResult.getFieldError().getDefaultMessage();
             return ResponseEntity.ok(error);
         } else {
-            return ResponseEntity.ok(sizeService.add(sizeRequest));
+            return ResponseEntity.ok(promotionService.add(promotionRequest));
         }
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> update(@RequestBody SizeRequest sizeRequest, @PathVariable Integer id) {
-        return ResponseEntity.ok(sizeService.update(sizeRequest, id));
+    public ResponseEntity<?> update(@RequestBody PromotionRequest promotionRequest, @PathVariable Integer id) {
+        return ResponseEntity.ok(promotionService.update(promotionRequest, id));
     }
 
     @PutMapping("/delete/{id}")
     public ResponseEntity<?> delete(@PathVariable Integer id) {
-        if (sizeService.delete(id)) {
+        if (promotionService.delete(id)) {
             return ResponseEntity.ok("Delete Success");
         } else {
             return ResponseEntity.ok("Delete Fail");
@@ -53,6 +51,6 @@ public class SizeController {
 
     @GetMapping("/search")
     public ResponseEntity<?> search(@RequestParam String name, @RequestParam(defaultValue = "0") int page) {
-        return ResponseEntity.ok(sizeService.search(name, page).getContent());
+        return ResponseEntity.ok(promotionService.search(name, page).getContent());
     }
 }
