+++ src/main/java/com/sd64/novastore/controller/admin/SizeDetailController.java
@@ -1,4 +1,4 @@
-package com.sd64.novastore.controller;
+package com.sd64.novastore.controller.admin;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
